Move constUtils
