+++ src/components/principal/Sobre.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { View, Text, StyleSheet, Image, TouchableOpacity, ScrollView } from 'react-native';
 import { logoAzUnB, logoAzProEpi } from '../../imgs/imageConst';
 import { scale } from '../../utils/scallingUtils';
-import { Redirect, textos } from '../../constUtils';
+import { Redirect, textos } from '../../utils/constUtils';
 import translate from "../../../locales/i18n";
 
 class Sobre extends Component {
