+++ src/components/telainicial/Login.js
@@ -5,7 +5,7 @@ import AwesomeAlert from 'react-native-awesome-alerts';
 import Emoji from 'react-native-emoji';
 import { scale } from '../../utils/scallingUtils';
 import translate from '../../../locales/i18n';
-import { API_URL } from '../../constUtils';
+import { API_URL } from '../../utils/constUtils';
 import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scrollview';
 
 class Login extends Component {
